requirements.txt
Flask~=3.1.0
Flask-Cors~=5.0.0
PyYAML~=6.0.2
langchain-community~=0.3.17
langchain-core~=0.3.34
langchain~=0.3.18
langchain-openai~=0.3.4
openai~=1.61.0
SQLAlchemy~=2.0.37
reportlab~=4.2.5
fastapi~=0.115.8
pydantic~=2.10.6
python-dotenv~=1.0.1
